Problem 3 now runs to completion in Haskell.
Memory requested too large error not present anymore.
To find if a number was prime it was finding all the factors and seeing
if the the length was 0 rather than short circuiting. Futher, use head
instead of last on the factors list and reverse the order we find them in
so that we stop once we find the largest.